+++ vars/terraform.groovy
@@ -23,7 +23,7 @@ void apply(body) {
       if (config.var_file) {
         assert fileExists(config.var_file) : "The var file ${config.var_file} does not exist!"
 
-        cmd += " -var_file=${config.var_file}"
+        cmd += " -var-file=${config.var_file}"
       }
       if (config.var) {
         assert (config.var instanceof Map) : 'The var parameter must be a Map.'
@@ -83,7 +83,7 @@ void destroy(body) {
     if (config.var_file) {
       assert fileExists(config.var_file) : "The var file ${config.var_file} does not exist!"
 
-      cmd += " -var_file=${config.var_file}"
+      cmd += " -var-file=${config.var_file}"
     }
     if (config.var) {
       assert (config.var instanceof Map) : 'The var parameter must be a Map.'
@@ -224,7 +224,7 @@ void imports(body) {
     if (config.var_file) {
       assert fileExists(config.var_file) : "The var file ${config.var_file} does not exist!"
 
-      cmd += " -var_file=${config.var_file}"
+      cmd += " -var-file=${config.var_file}"
     }
     if (config.var) {
       assert (config.var instanceof Map) : 'The var parameter must be a Map.'
@@ -352,7 +352,7 @@ def plan(body) {
     if (config.var_file) {
       assert fileExists(config.var_file) : "The var file ${config.var_file} does not exist!"
 
-      cmd += " -var_file=${config.var_file}"
+      cmd += " -var-file=${config.var_file}"
     }
     if (config.var) {
       assert (config.var instanceof Map) : 'The var parameter must be a Map.'