+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Add `refreshOnly` parameter to `plan` method.
 - Add `display` parameter to `output` method.
 - Add `backendConfig` parameter to `init` method.
+- Fix `var_file` parameter in all relevant methods.
 
 ### 1.6.0
 **Helm**